Move worker out to its own file
