+++ lib/flipper/adapters/http_read_async.rb
@@ -1,6 +1,8 @@
 require 'flipper'
 require 'flipper/adapters/http'
 require 'flipper/adapters/memory'
+require 'flipper/adapters/http_read_async/worker'
+require 'thread'
 
 module Flipper
   module Adapters
@@ -9,99 +11,30 @@ module Flipper
 
       attr_reader :name
 
-      require 'thread'
-      require 'logger'
-
-      class Worker
-        PREFIX = "[flipper http read async adapter]".freeze
-
-        attr_reader :thread, :pid, :mutex, :logger, :interval
-
-        def initialize(options = {}, &block)
-          @thread = nil
-          @pid = Process.pid
-          @mutex = Mutex.new
-          @logger = options.fetch(:logger) { Logger.new(STDOUT) }
-          @interval = options.fetch(:interval, 10)
-          @start_automatically = options.fetch(:start_automatically, true)
-          @block = block
-
-          if options.fetch(:shutdown_automatically, true)
-            at_exit { stop }
-          end
-        end
-
-        def start
-          reset if forked?
-          ensure_worker_running
-        end
-
-        def stop
-          logger.info { "#{PREFIX} Stopping worker" }
-          @thread&.kill
-        end
-
-        def run
-          loop do
-            logger.info { "#{PREFIX} Sleeping for #{interval} seconds" }
-            sleep interval
-
-            begin
-              logger.info { "#{PREFIX} Making a checkity checkity" }
-              @block.call
-            rescue => exception
-              # log this or something
-            end
-          end
-        end
-
-        private
-
-        def forked?
-          pid != Process.pid
-        end
-
-        def ensure_worker_running
-          # Return early if thread is alive and avoid the mutex lock and unlock.
-          return if thread_alive?
-
-          # If another thread is starting worker thread, then return early so this
-          # thread can enqueue and move on with life.
-          return unless mutex.try_lock
-
-          begin
-            return if thread_alive?
-            @thread = Thread.new { run }
-            logger.debug { "#{PREFIX} Worker thread [#{@thread.object_id}] started" }
-          ensure
-            mutex.unlock
-          end
-        end
-
-        def thread_alive?
-          @thread && @thread.alive?
-        end
-
-        def reset
-          @pid = Process.pid
-          mutex.unlock if mutex.locked?
-        end
-      end
-
       def initialize(options = {})
+        @name = :http_async
         @memory_adapter = Memory.new
+        @mutex = Mutex.new
+
+        # let's not start cold since cloud is configured with a local adapter,
+        # instead we can initially populate the memory adapter with the
+        # local adapter
         if adapter = options[:start_with]
           @memory_adapter.import(adapter)
         end
+
         @http_adapter = Http.new(options)
-        @mutex = Mutex.new
-        @name = :http_async
-        @worker = Worker.new(options[:worker] || {}) {
+        @interval = options.fetch(:interval, 10)
+
+        # setup the worker and default interval to top level interval if not
+        # set, setting interval in worker options takes precendence
+        worker_options = options[:worker] || {}
+        worker_options[:interval] ||= @interval
+        @worker = Worker.new(worker_options) {
           adapter = Memory.new
           adapter.import(@http_adapter)
-          @mutex.synchronize {
-            @memory_adapter.import(adapter)
-          }
+          # lock when updating the memory adapter in the main thread
+          @mutex.synchronize { @memory_adapter.import(adapter) }
         }
         @worker.start
       end
